+++ templates/nextjs-boilerplate-template/src/app/layout.tsx
@@ -1,6 +1,7 @@
 import { ReactQueryProvider } from "@/components/ReactQueryProvider";
 import { WalletProvider } from "@/components/WalletProvider";
 import { Toaster } from "@/components/ui/toaster";
+import { WrongNetwork } from "@/components/WrongNetwork";
 import type { Metadata } from "next";
 import type { ReactNode } from "react";
 import "./globals.css";
@@ -21,6 +22,7 @@ export default function RootLayout({
         <WalletProvider>
           <ReactQueryProvider>
             <div id="root">{children}</div>
+            <WrongNetwork />
             <Toaster />
           </ReactQueryProvider>
         </WalletProvider>