+++ templates/nft-minting-dapp-template/frontend/main.tsx
@@ -8,6 +8,7 @@ import App from "@/App";
 import { WalletProvider } from "@/components/WalletProvider";
 import { Toaster } from "@/components/ui/toaster";
 import { TooltipProvider } from "@/components/ui/tooltip";
+import { WrongNetwork } from "@/components/WrongNetwork";
 
 const queryClient = new QueryClient();
 
@@ -17,6 +18,7 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
       <QueryClientProvider client={queryClient}>
         <TooltipProvider delayDuration={100}>
           <App />
+          <WrongNetwork />
           <Toaster />
         </TooltipProvider>
       </QueryClientProvider>